requirements.txt
pyaml
rospkg
numpy
opencv-contrib-python
scipy
pynput
pybind11
pybind11-global
scipy
pynput
whisper
sounddevice
google
anthropic
dynamixel-sdk
